requirements.txt
# ── Core ────────────────────────────────────────────────────────────────────
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.2.3

# ── Storage ─────────────────────────────────────────────────────────────────
google-cloud-bigquery==3.17.2
google-cloud-bigquery-storage==2.24.0
db-dtypes==1.2.0

# ── LLM / Agent ─────────────────────────────────────────────────────────────
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.36
langchain-core==0.1.46
openai==1.23.6
anthropic==0.23.1

# ── Dashboard ───────────────────────────────────────────────────────────────
streamlit==1.33.0
plotly==5.21.0

# ── Data / Transform ────────────────────────────────────────────────────────
scipy==1.13.0
scikit-learn==1.4.2

# ── Orchestration ───────────────────────────────────────────────────────────
apache-airflow==2.9.0

# ── Testing ─────────────────────────────────────────────────────────────────
pytest==8.1.1
pytest-mock==3.14.0

# ── Dev / CLI ───────────────────────────────────────────────────────────────
rich==13.7.1
